+++ changelog.md
@@ -1,2 +1,2 @@
 # Fabric
-- fix leaf decay with FallingTree
+- add compat with kleeslabs